Module B: drop delay in WAIT if switching to RUN
Drop `delay(1100)` call if we're switching to RUN. This removes the
unnecessary wait time during which module A probably dissynchronises
counter with module B.

Additional change: bigger delay (1000→1100). This makes sure we get more
messages from module A (it may be a not important change since flushing
was introduced in parent commit, but I like it anyway).